Donar a base de datos

+++ app/src/main/java/mx/itesm/dibujandounmaana/view/Donar.kt
@@ -83,7 +83,7 @@ class Donar : Fragment() {
 
                     //Enviar Donacion
                     val nuevaDonacion= Donar(binding.spinner2.selectedItem.toString(),
-                        binding.etCantidad.text.toString())
+                        binding.etCantidad.text.toString().toInt())
                     viewModel.enviarDonacion(nuevaDonacion)
                 }
             },
